rename files and vars for clarity
