+++ lib/pageSlug.js
@@ -7,7 +7,7 @@
   Because content relationships in Prismic are limited to one-level deep, we need to use array of parent pages instead of a single parent page
 */
 import { pageFetchLinks } from '../constants/page'
-import { PrismicClient } from './api'
+import { client } from './prismic'
 import asyncForEach from './asyncForEach'
 
 export function getPageSlug(page) {
@@ -85,7 +85,7 @@ export async function fetchParentPageLinks(obj) {
       newDoc = docRecords[doc.id]
     } else {
       // only look up this record once
-      newDoc = docRecords[doc.id] = await PrismicClient.getByID(doc.id, {
+      newDoc = docRecords[doc.id] = await client.getByID(doc.id, {
         fetchLinks,
       })
     }
