+++ lib/pathFormation.js
@@ -1,5 +1,5 @@
 import * as prismic from '@prismicio/client'
-import { PrismicClient } from 'lib/api'
+import { client } from 'lib/prismic'
 import { pageFetchLinks } from 'constants/page'
 import { getPageSlug } from './pageSlug'
 
@@ -14,7 +14,7 @@ export async function getPagePaths({ pageType = 'page', excludeId }) {
   let allPages = []
 
   while (morePages) {
-    const { results, total_pages: totalPages, page } = await PrismicClient.get({
+    const { results, total_pages: totalPages, page } = await client.get({
       predicates: [prismic.Predicates.at('document.type', pageType)],
       pageSize,
       page: curPage,
