+++ lib/prismic.js
@@ -9,6 +9,6 @@ export const api = {
   LOCALE: process.env.PRISMIC_REPOSITORY_LOCALE,
 }
 
-export const PrismicClient = prismic.createClient(api.ENDPOINT, {
+export const client = prismic.createClient(api.ENDPOINT, {
   accessToken: api.TOKEN,
 })
